backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
CacheControl==0.14.4
cachetools==6.2.4
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
claude-agent-sdk==0.1.5
click==8.3.0
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.124.4
firebase_admin==7.1.0
google-api-core==2.28.1
google-auth==2.45.0
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.7.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lit==18.1.8
markdown-it-py==4.0.0
mcp==1.19.0
mdurl==0.1.2
msgpack==1.1.2
proto-plus==1.27.0
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pywatchman==1.4.1
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.38.0
websockets==15.0.1
